Exc 4.21

+++ controllers/blogs.js
@@ -37,8 +37,19 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    return response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'token invalid' })
+    }
 })
 
 blogRouter.put('/:id', async (request, response) => {
